scripts/requirements.txt
portia-sdk-python>=0.7.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
asyncio
typing
datetime
